Extracted Covariance class
